# Conflicts:
#	core/src/main/java/fr/uge/modules/linking/Linking.java

+++ core/src/main/java/fr/uge/modules/api/model/linking/Link.java
@@ -1,4 +1,6 @@
 package fr.uge.modules.api.model.linking;
 
-public record Link(long id_source, long id_target, float proximity) {
+import java.math.BigDecimal;
+
+public record Link(long id_source, long id_target, BigDecimal proximity) {
 }
